# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ContentKit",
    "version": "v0",
    "description": "ContentKit's HTTP API, from the prefix the host mounts contentkit.Runtime.Handler at. Each module sits at its own sub-path (x-contentkit-modules); a host that mounts a module alone serves its routes beneath that mount instead. Generated from the route catalog (internal/httpapi) by `go generate ./internal/contract`; do not edit."
  },
  "paths": {
...
    "/taxonomy/nodes": {
      "get": {
        "tags": [
          "taxonomy"
        ],
        "summary": "Lists nodes: filtered, ordered and paged.",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "description": "only nodes of this kind",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "description": "only nodes in these states; active by default",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "id",
            "in": "query",
            "description": "only these taxonomy ids",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "slug",
            "in": "query",
            "description": "only the node with this slug",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "in": "query",
            "description": "the language names resolve in",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language_mode",
            "in": "query",
            "description": "what a node without a name in language gets: fallback (the default), strict or required",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name_prefix",
            "in": "query",
            "description": "names starting with this",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "names containing this",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "content_kind",
            "in": "query",
            "description": "count contents of this kind",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "min_count",
            "in": "query",
            "description": "only nodes with at least this many contents",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "name, count, created, oldest or updated; taxonomy id by default",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "the previous page's next_cursor (default order only)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "nodes to skip",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "page size",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodePage"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "host": []
          }
        ],
        "x-contentkit-module": "taxonomy",
        "x-contentkit-auth": "staff",
        "x-contentkit-permission": "Taxonomy",
        "x-contentkit-errors": [],
        "x-contentkit-error-sets": [
          "module:taxonomy",
          "tier:staff",
          "request"
        ]
      },
      "post": {
        "tags": [
          "taxonomy"
        ],
        "summary": "Creates nodes with their names.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/NodeInput"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Node"
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "host": []
          }
        ],
        "x-contentkit-module": "taxonomy",
        "x-contentkit-auth": "staff",
        "x-contentkit-permission": "Taxonomy",
        "x-contentkit-errors": [
          "conflict"
        ],
        "x-contentkit-error-sets": [
          "module:taxonomy",
          "tier:staff",
          "request"
        ]
      }
    },
    "/taxonomy/nodes/{id}": {
      "get": {
        "tags": [
          "taxonomy"
        ],
        "summary": "A node with its names and edges.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeDetail"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "host": []
          }
        ],
        "x-contentkit-module": "taxonomy",
        "x-contentkit-auth": "staff",
        "x-contentkit-permission": "Taxonomy",
        "x-contentkit-errors": [
          "not_found"
        ],
        "x-contentkit-error-sets": [
          "module:taxonomy",
          "tier:staff",
          "request"
        ]
      },
      "patch": {
        "tags": [
          "taxonomy"
        ],
        "summary": "Updates a node's given fields.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NodeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Node"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "host": []
          }
        ],
        "x-contentkit-module": "taxonomy",
        "x-contentkit-auth": "staff",
        "x-contentkit-permission": "Taxonomy",
        "x-contentkit-errors": [
          "conflict",
          "not_found"
        ],
        "x-contentkit-error-sets": [
          "module:taxonomy",
          "tier:staff",
          "request"
        ]
      },
      "delete": {
        "tags": [
          "taxonomy"
        ],
        "summary": "Marks a node deleted.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Node"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "host": []
          }
        ],
        "x-contentkit-module": "taxonomy",
        "x-contentkit-auth": "staff",
        "x-contentkit-permission": "Taxonomy",
        "x-contentkit-errors": [
          "not_found"
        ],
        "x-contentkit-error-sets": [
          "module:taxonomy",
          "tier:staff",
          "request"
        ]
      }
    },
    "/taxonomy/nodes/{id}/names": {
      "put": {
        "tags": [
          "taxonomy"
        ],
        "summary": "Replaces a node's names.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Name"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeDetail"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "host": []
          }
        ],
        "x-contentkit-module": "taxonomy",
        "x-contentkit-auth": "staff",
        "x-contentkit-permission": "Taxonomy",
        "x-contentkit-errors": [
          "conflict",
          "not_found"
        ],
        "x-contentkit-error-sets": [
          "module:taxonomy",
          "tier:staff",
          "request"
        ]
      },
      "post": {
        "tags": [
          "taxonomy"
        ],
        "summary": "Adds names to a node.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Name"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeDetail"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "host": []
          }
        ],
        "x-contentkit-module": "taxonomy",
        "x-contentkit-auth": "staff",
        "x-contentkit-permission": "Taxonomy",
        "x-contentkit-errors": [
          "conflict",
          "not_found"
        ],
        "x-contentkit-error-sets": [
          "module:taxonomy",
          "tier:staff",
          "request"
        ]
      },
      "delete": {
        "tags": [
          "taxonomy"
        ],
        "summary": "Removes names from a node.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Name"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeDetail"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "host": []
          }
        ],
        "x-contentkit-module": "taxonomy",
        "x-contentkit-auth": "staff",
        "x-contentkit-permission": "Taxonomy",
        "x-contentkit-errors": [
          "not_found"
        ],
        "x-contentkit-error-sets": [
          "module:taxonomy",
          "tier:staff",
          "request"
        ]
      }
    },
    "/taxonomy/nodes/{id}/merge": {
      "post": {
        "tags": [
          "taxonomy"
        ],
        "summary": "Merges a node into another: its assignments, names and edges move, and its URL redirects.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MergeReport"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "host": []
          }
        ],
        "x-contentkit-module": "taxonomy",
        "x-contentkit-auth": "staff",
        "x-contentkit-permission": "Taxonomy",
        "x-contentkit-errors": [
          "conflict",
          "not_found"
        ],
        "x-contentkit-error-sets": [
          "module:taxonomy",
          "tier:staff",
          "request"
        ]
      }
    },
    "/taxonomy/edges": {
      "post": {
        "tags": [
          "taxonomy"
        ],
        "summary": "Adds edges.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Edge"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EdgesWritten"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "host": []
          }
        ],
        "x-contentkit-module": "taxonomy",
        "x-contentkit-auth": "staff",
        "x-contentkit-permission": "Taxonomy",
        "x-contentkit-errors": [
          "conflict",
          "not_found"
        ],
        "x-contentkit-error-sets": [
          "module:taxonomy",
          "tier:staff",
          "request"
        ]
      },
      "delete": {
        "tags": [
          "taxonomy"
        ],
        "summary": "Removes edges.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Edge"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EdgesWritten"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "host": []
          }
        ],
        "x-contentkit-module": "taxonomy",
        "x-contentkit-auth": "staff",
        "x-contentkit-permission": "Taxonomy",
        "x-contentkit-errors": [],
        "x-contentkit-error-sets": [
          "module:taxonomy",
          "tier:staff",
          "request"
        ]
      }
    },
...
    "/taxonomy/effective": {
      "post": {
        "tags": [
          "taxonomy"
        ],
        "summary": "The effective tags (work and version) of each content.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ContentRef"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EffectiveTagsOf"
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "host": []
          }
        ],
        "x-contentkit-module": "taxonomy",
        "x-contentkit-auth": "staff",
        "x-contentkit-permission": "Taxonomy",
        "x-contentkit-errors": [],
        "x-contentkit-error-sets": [
          "module:taxonomy",
          "tier:staff",
          "request"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MergeInput": {
        "type": "object",
        "properties": {
          "into_taxonomy_id": {
            "type": "string"
          }
        }
      },
...
      "NodeInput": {
        "type": "object",
        "properties": {
          "taxonomy_id": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "names": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Name"
            }
          },
          "source_revision": {
            "type": "integer"
          }
        }
      },
      "NodePage": {
        "type": "object",
        "properties": {
          "nodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NodeRow"
            }
          },
          "next_cursor": {
            "type": "string"
          },
          "total": {
            "type": "integer"
          }
        },
        "required": [
          "nodes",
          "total"
        ]
      },
...
      "NodeUpdate": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "enum": [
              "active",
              "deleted",
              "merged"
            ]
          },
          "source_revision": {
            "type": "integer"
          }
        }
      },
...
    },
    "responses": {
      "Error": {
        "description": "A refusal: a code of the operation's x-contentkit-errors or of an x-contentkit-error-sets set it names.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorReply"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "host": {
        "type": "http",
        "scheme": "bearer",
        "description": "Whatever the host's auth middleware accepts: ContentKit reads the actor it puts in the request context."
      }
    }
  },
...
}